+++ .prettierrc
@@ -1,8 +1,13 @@
 {
-  "printWidth": 85,
   "semi": true,
   "singleQuote": true,
-  "tabWidth": 2,
+  "printWidth": 120,
   "trailingComma": "es5",
-  "arrowParens": "always"
+  "arrowParens": "always",
+  "tabWidth": 2,
+  "useTabs": false,
+  "quoteProps": "as-needed",
+  "jsxSingleQuote": true,
+  "bracketSpacing": true,
+  "bracketSameLine": true
 }
