--- jsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["resources/js/*"]
-    }
-  },
-  "exclude": ["node_modules", "public"]
-}
