+++ package.json
@@ -1,35 +1,41 @@
 {
-  "private": true,
-  "scripts": {
-    "prepare": "husky install && chmod ug+x .husky/*",
-    "dev": "vite",
-    "build": "vite build",
-    "format": "prettier 'resources/{js,jsx}/**/*.{js,jsx}' --write",
-    "pre-commit": "npm run format && pretty-quick --staged"
-  },
-  "devDependencies": {
-    "@emotion/react": "^11.10.6",
-    "@inertiajs/react": "^1.0.2",
-    "@mantine/core": "^6.0.5",
-    "@mantine/hooks": "^6.0.5",
-    "@mantine/modals": "^6.0.5",
-    "@tabler/icons": "^1.117.0",
-    "@vitejs/plugin-react": "^3.0.1",
-    "autoprefixer": "^10.4.2",
-    "axios": "^1.2.2",
-    "classnames": "^2.3.2",
-    "husky": "^8.0.2",
-    "laravel-vite-plugin": "^0.7.3",
-    "lodash": "^4.17.19",
-    "postcss": "^8.4.6",
-    "postcss-nesting": "^10.2.0",
-    "prettier": "^2.8.0",
-    "pretty-quick": "^3.1.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-hot-toast": "^2.4.0",
-    "react-use": "^17.4.0",
-    "tailwindcss": "^3.1.0",
-    "vite": "^4.0.4"
-  }
+    "private": true,
+    "scripts": {
+        "prepare": "husky install && chmod ug+x .husky/*",
+        "pre-commit": "npm run format && pretty-quick --staged",
+        "format": "prettier 'resources/{js,jsx}/**/*.{js,jsx}' --write",
+        "dev": "vite",
+        "build": "vite build"
+    },
+    "devDependencies": {
+        "@emotion/react": "^11.10.6",
+        "@inertiajs/react": "^1.0.2",
+        "@mantine/core": "^6.0.5",
+        "@mantine/hooks": "^6.0.5",
+        "@mantine/modals": "^6.0.5",
+        "@tabler/icons": "^1.117.0",
+        "@tabler/icons-react": "^2.18.0",
+        "@types/lodash": "^4.14.194",
+        "@types/react": "^18.2.6",
+        "@types/react-dom": "^18.2.4",
+        "@vitejs/plugin-react": "^3.0.1",
+        "autoprefixer": "^10.4.2",
+        "axios": "^1.2.2",
+        "classnames": "^2.3.2",
+        "husky": "^8.0.2",
+        "laravel-vite-plugin": "^0.7.3",
+        "lodash": "^4.17.19",
+        "postcss": "^8.4.6",
+        "postcss-nesting": "^10.2.0",
+        "prettier": "^2.8.0",
+        "pretty-quick": "^3.1.3",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-hot-toast": "^2.4.0",
+        "react-use": "^17.4.0",
+        "tailwindcss": "^3.1.0",
+        "ts-loader": "^9.4.2",
+        "typescript": "^5.0.4",
+        "vite": "^4.0.4"
+    }
 }
