+++ resources/js/Components/DeleteButton.tsx
@@ -1,0 +1,5 @@
+export default ({ onDelete, children }) => (
+    <button className='text-red-600 focus:outline-none hover:underline' tabIndex='-1' type='button' onClick={onDelete}>
+        {children}
+    </button>
+);
