+++ resources/js/Components/FlashErrors.tsx
