+++ resources/js/Components/ResponsivePagination.tsx
