+++ resources/js/Layouts/AuthenticatedLayout.tsx
