+++ resources/js/Pages/Auth/ConfirmPassword.tsx
