+++ resources/js/Pages/Auth/ForgotPassword.tsx
