+++ resources/js/Pages/Auth/Register.tsx
