+++ resources/js/Pages/Auth/VerifyEmail.tsx
