+++ resources/js/Pages/Management/Permissions/Create.tsx
