+++ tailwind.config.js
@@ -6,7 +6,7 @@ module.exports = {
     './vendor/laravel/framework/src/Illuminate/Pagination/resources/views/*.blade.php',
     './storage/framework/views/*.php',
     './resources/views/**/*.blade.php',
-    './resources/js/**/*.jsx',
+    './resources/js/**/*.tsx',
   ],
 
   theme: {
